# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ModelKit API",
        "title": "ModelKit API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/model": {
            "get": {
                "description": "获取模型列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "获取模型列表",
                "operationId": "list-model",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.AllModelResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "更新模型",
                "operationId": "update-model",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Model"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "创建模型",
                "operationId": "create-model",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Model"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "删除模型",
                "operationId": "delete-model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        },
        "/api/v1/model/check": {
            "post": {
                "description": "检查模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "检查模型",
                "operationId": "check-model",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CheckModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Model"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/model/provider/supported": {
            "get": {
                "description": "获取供应商支持的模型列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "获取供应商支持的模型列表",
                "operationId": "get-provider-model-list",
                "parameters": [
                    {
                        "type": "string",
                        "name": "api_header",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "api_key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "base_url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "SiliconFlow",
                            "OpenAI",
                            "Ollama",
                            "DeepSeek",
                            "Moonshot",
                            "AzureOpenAI",
                            "BaiZhiCloud",
                            "Hunyuan",
                            "BaiLian",
                            "Volcengine"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "ModelProviderSiliconFlow",
                            "ModelProviderOpenAI",
                            "ModelProviderOllama",
                            "ModelProviderDeepSeek",
                            "ModelProviderMoonshot",
                            "ModelProviderAzureOpenAI",
                            "ModelProviderBaiZhiCloud",
                            "ModelProviderHunyuan",
                            "ModelProviderBaiLian",
                            "ModelProviderVolcengine"
                        ],
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "llm",
                            "coder",
                            "embedding",
                            "audio",
                            "reranker"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "ModelTypeLLM",
                            "ModelTypeCoder",
                            "ModelTypeEmbedding",
                            "ModelTypeAudio",
                            "ModelTypeReranker"
                        ],
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.GetProviderModelListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "consts.ModelProvider": {
            "type": "string",
            "enum": [
                "SiliconFlow",
                "OpenAI",
                "Ollama",
                "DeepSeek",
                "Moonshot",
                "AzureOpenAI",
                "BaiZhiCloud",
                "Hunyuan",
                "BaiLian",
                "Volcengine"
            ],
            "x-enum-varnames": [
                "ModelProviderSiliconFlow",
                "ModelProviderOpenAI",
                "ModelProviderOllama",
                "ModelProviderDeepSeek",
                "ModelProviderMoonshot",
                "ModelProviderAzureOpenAI",
                "ModelProviderBaiZhiCloud",
                "ModelProviderHunyuan",
                "ModelProviderBaiLian",
                "ModelProviderVolcengine"
            ]
        },
        "consts.ModelStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive"
            ],
            "x-enum-varnames": [
                "ModelStatusActive",
                "ModelStatusInactive"
            ]
        },
        "consts.ModelType": {
            "type": "string",
            "enum": [
                "llm",
                "coder",
                "embedding",
                "audio",
                "reranker"
            ],
            "x-enum-varnames": [
                "ModelTypeLLM",
                "ModelTypeCoder",
                "ModelTypeEmbedding",
                "ModelTypeAudio",
                "ModelTypeReranker"
            ]
        },
...
        "domain.AllModelResp": {
            "type": "object",
            "properties": {
                "providers": {
                    "description": "提供商列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ProviderModel"
                    }
                }
            }
        },
...
        "domain.CheckModelReq": {
            "type": "object",
            "required": [
                "api_base",
                "api_key",
                "model_name",
                "provider",
                "type"
            ],
            "properties": {
                "api_base": {
                    "description": "接口地址",
                    "type": "string"
                },
                "api_header": {
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥",
                    "type": "string"
                },
                "api_version": {
                    "type": "string"
                },
                "model_name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "provider": {
                    "description": "提供商",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelProvider"
                        }
                    ]
                },
                "type": {
                    "enum": [
                        "llm",
                        "coder",
                        "embedding",
                        "rerank"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelType"
                        }
                    ]
                }
            }
        },
...
        "domain.CreateModelReq": {
            "type": "object",
            "required": [
                "api_base",
                "model_name",
                "provider"
            ],
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "api_header": {
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥 如：sk-xxxx",
                    "type": "string"
                },
                "api_version": {
                    "type": "string"
                },
                "model_name": {
                    "description": "模型名称 如: deepseek-v3",
                    "type": "string"
                },
                "model_type": {
                    "description": "模型类型 llm:对话模型 coder:代码模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelType"
                        }
                    ]
                },
                "param": {
                    "description": "高级参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ModelParam"
                        }
                    ]
                },
                "provider": {
                    "description": "提供商",
                    "enum": [
                        "SiliconFlow",
                        "OpenAI",
                        "Ollama",
                        "DeepSeek",
                        "Moonshot",
                        "AzureOpenAI",
                        "BaiZhiCloud",
                        "Hunyuan",
                        "BaiLian",
                        "Volcengine",
                        "Other"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelProvider"
                        }
                    ]
                },
                "show_name": {
                    "description": "模型显示名称",
                    "type": "string"
                }
            }
        },
...
        "domain.GetProviderModelListResp": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ProviderModelListItem"
                    }
                }
            }
        },
...
        "domain.Model": {
            "type": "object",
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "api_header": {
                    "description": "接口头 如：Authorization: Bearer sk-xxxx",
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥 如：sk-xxxx",
                    "type": "string"
                },
                "api_version": {
                    "description": "接口版本 如：2023-05-15",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "id": {
                    "description": "模型ID",
                    "type": "string"
                },
                "input": {
                    "description": "输入token数",
                    "type": "integer"
                },
                "is_active": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "is_internal": {
                    "description": "是否内部模型",
                    "type": "boolean"
                },
                "model_name": {
                    "description": "模型名称 如: deepseek-v3",
                    "type": "string"
                },
                "model_type": {
                    "description": "模型类型 llm:对话模型 coder:代码模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelType"
                        }
                    ]
                },
                "output": {
                    "description": "输出token数",
                    "type": "integer"
                },
                "param": {
                    "description": "高级参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ModelParam"
                        }
                    ]
                },
                "provider": {
                    "description": "提供商",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelProvider"
                        }
                    ]
                },
                "show_name": {
                    "description": "模型显示名称",
                    "type": "string"
                },
                "status": {
                    "description": "状态 active:启用 inactive:禁用",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelStatus"
                        }
                    ]
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "integer"
                }
            }
        },
        "domain.ModelBasic": {
            "type": "object",
            "required": [
                "provider"
            ],
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "provider": {
                    "description": "提供商",
                    "enum": [
                        "SiliconFlow",
                        "OpenAI",
                        "Ollama",
                        "DeepSeek",
                        "Moonshot",
                        "AzureOpenAI",
                        "BaiZhiCloud",
                        "Hunyuan",
                        "BaiLian",
                        "Volcengine",
                        "Other"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelProvider"
                        }
                    ]
                }
            }
        },
...
        "domain.ModelParam": {
            "type": "object",
            "properties": {
                "context_window": {
                    "type": "integer"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "r1_enabled": {
                    "type": "boolean"
                },
                "support_computer_use": {
                    "type": "boolean"
                },
                "support_images": {
                    "type": "boolean"
                },
                "support_prompt_cache": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.ProviderModel": {
            "type": "object",
            "properties": {
                "models": {
                    "description": "模型列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ModelBasic"
                    }
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                }
            }
        },
        "domain.ProviderModelListItem": {
            "type": "object",
            "properties": {
                "model": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateModelReq": {
            "type": "object",
            "required": [
                "provider"
            ],
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "api_header": {
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥 如：sk-xxxx",
                    "type": "string"
                },
                "api_version": {
                    "type": "string"
                },
                "id": {
                    "description": "模型ID",
                    "type": "string"
                },
                "model_name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "param": {
                    "description": "高级参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ModelParam"
                        }
                    ]
                },
                "provider": {
                    "description": "提供商",
                    "enum": [
                        "SiliconFlow",
                        "OpenAI",
                        "Ollama",
                        "DeepSeek",
                        "Moonshot",
                        "AzureOpenAI",
                        "BaiZhiCloud",
                        "Hunyuan",
                        "BaiLian",
                        "Volcengine",
                        "Other"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelProvider"
                        }
                    ]
                },
                "show_name": {
                    "description": "模型显示名称",
                    "type": "string"
                },
                "status": {
                    "description": "状态 active:启用 inactive:禁用",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelStatus"
                        }
                    ]
                }
            }
        },
...
        "web.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}